Implement (incomplete) dependency checking.
Targets still do not depend on .do-files, and a lot of the edge cases
are still not handled correctly. Furthermore some error-checking code is
still missing, which could possibly crash the program (partially marked
with comments), as well as some free() calls.

An utitlity python script (print_dep.py) was also added to aid in
debugging matters.