Make targets depend on do-files.
This includes two different scenarios, first if the do-files which was
used to build the target has changed the target must be rebuilt.

Second, if a target was build by a lower priority do-file, like
default.fuu.do, then it must be rebuilt if <target>.fuu.do appears.

Note that at this point, Redofile as do-file is semi-supported, as it's
future is uncertain.