Replace current tests with a hole new test suite
This new test suite is based on just on do scripts, much like apenwarr's redo
implementation has done it. This makes tests both simpler and free from any
additional dependencies. The naming was also improved a lot.

There are also a lot of new tests, and many more to be implemeted later.